Fix the logic for the default fallback sdk version in ld64
The apple case for the default uses sysctl to get the kernel
version, which is a version number like 18.7.0, where the major
version, minus 4, is used for setting the minor version of
the macOS version number, translating 18.7.0 into 10.14.

The previous ld64-port specific case of trying to hardcode 10.5,
just ends up parsed as kernel version 10, i.e. 10.6, regardless of
which version the default string is set to.

Instead of passing the cctools/ld64-port specific default to the
kernel version parsing logic, just set the result of
parseVersionNumber32 as the version number.

Change the default to 10.6, which is what actually was set before
in practice.